pyproject.toml
[project]
name = "ai-doubt-resolver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deep-translator==1.11.4",
    "euriai>=0.3.30",
    "faiss-cpu==1.11.0",
    "langchain==0.2.17",
    "langchain-community==0.2.19",
    "numpy>=1.26.0,<2.0.0",
    "pymupdf==1.26.1",
    "pypdf2==3.0.1",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "streamlit==1.35.0",
]
